refactor: split src/main.rs into cli/, app/, direct.rs
main.rs was 3089 lines mixing CLI parsing, subcommand handlers, the
foreground event loop, pane lifecycle helpers, render orchestration,
and the IPC dispatch table. Splitting along concern boundaries makes
each chunk independently reviewable and unblocks the planned daemon
refactor (#17).

New layout:
- src/main.rs          (143 lines, dispatcher only)
- src/direct.rs        (--no-daemon entry: terminal setup + teardown)
- src/cli/{mod,parse,help}.rs
- src/app/state.rs        (InputMode, TextSelection, DragState,
                           RenderUpdate, SnapshotExtra)
- src/app/render_ctl.rs   (frame composition + dirty-set helpers)
- src/app/lifecycle.rs    (spawn/split/replace/resize/snapshot/clipboard)
- src/app/bootstrap.rs    (build_initial_state + Procfile + launches)
- src/app/attach.rs       (cmd_ls/kill/rename/attach/init/from/doctor)
- src/app/input_dispatch.rs (handle_ipc_command)
- src/app/event_loop.rs   (the run() loop)

server.rs swapped its `super::*` references for the new module paths.
CONTRIBUTING.md gained a "Module anatomy" section so new contributors
have a map.

Behavior is unchanged: every moved item kept its visibility and
signature. 105 unit + 4 integration tests pass; render_hotpaths bench
shows no change (it pulls source files via #[path], not via the bin
crate, so the structural split can't affect it).

Spec deviation: app/event_loop.rs is 1249 lines, above the 500-line
target. Splitting run() further requires either a context struct or
many out-params for ~15 interleaved locals, which would change borrow
patterns and risk the "zero behavior change" gate. Tracked as
follow-up under the same issue.

Refs #23
